[Components] Cria componente de botão Fixes #1
