+++ src/components/atoms/index.export.ts
@@ -1,0 +1,5 @@
+import { Button} from './Button'
+
+export const atoms = {
+    Button,
+}
