+++ src/components/organisms/PageStatic/Page.stories.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Story, Meta } from '@storybook/react/types-6-0';
 
-import { Page, PageProps } from './Page';
-import * as HeaderStories from './Header.stories';
+import { Page, PageProps } from '../PageStatic/Page';
+import * as HeaderStories from '../../molecules/Header/Header.stories';
 
 export default {
-  title: 'Example/Page',
+  title: 'Organisms/Page',
   component: Page,
 } as Meta;
